pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proc2d"
version = "0.1.0"
description = "2D semiconductor process cross-section simulator (MVP)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "PyYAML>=6.0",
  "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "mypy>=1.10",
  "types-PyYAML>=6.0.12",
]
gui = ["streamlit>=1.35"]

[project.scripts]
proc2d = "proc2d.cli:main"
proc2d-gui = "proc2d.gui:main"

[tool.setuptools.packages.find]
include = ["proc2d*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --strict-markers"
markers = [
  "unit: fast isolated unit tests",
  "integration: cross-module integration tests",
  "adapter: CLI/GUI adapter tests",
  "slow: slow-running tests",
]

[tool.mypy]
python_version = "3.10"
files = [
  "proc2d/config",
  "proc2d/pipeline",
  "proc2d/services",
  "proc2d/analysis",
  "proc2d/app",
  "proc2d/cli.py",
  "proc2d/deck.py",
  "proc2d/gui_app.py",
]
pretty = true
show_error_codes = true
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["streamlit", "streamlit.*"]
ignore_missing_imports = true
